update rename files
