+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import 'package:repo_searcher/providers/homeProviders.dart';
+import 'package:repo_searcher/providers/home_providers.dart';
 
 import 'package:repo_searcher/pages/home.dart';
 import 'package:repo_searcher/pages/settings.dart';
