+++ lib/modules/item_pagenation.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:repo_searcher/providers/homeProviders.dart';
+import 'package:repo_searcher/providers/home_providers.dart';
 
 PagenationList(WidgetRef ref, int nowPage, int repoPerPage, int totalRepos){
